backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
google-generativeai==0.7.1
pydantic-settings==2.3.3
